server/requirements.txt
fastapi
uvicorn
pydantic
httpx
aiohttp
asyncio
faiss-cpu
sentence-transformers
openai
python-dotenv
beautifulsoup4
lxml
requests
pandas
numpy
chromadb
playwright
selenium
